Fix signing of query parameters
Previously, query parameters in the URL would not be encoded properly.
This change adds support for encoding query paramters in the request
`:url`, as well as in `:query-params`

    (one/sign-request consumer
      {:url "https://a.b/?foo=bar"
       :query-params {"baz" "qux"})

In the above example, both the query parameters in the URL, and the
query parameters in the map will be merged and signed.